// lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  contactEmail?: string | null;
}

export type JobWorkType = typeof JobWorkType[keyof typeof JobWorkType];


export const JobWorkType = {
  Remote: 'Remote',
  Hybrid: 'Hybrid',
  'On-site': 'On-site',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  Pending: 'Pending',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;

export interface Job {
  id: number;
  title: string;
  company: Company;
  location: string;
  workType: JobWorkType;
  /** @nullable */
  salary?: string | null;
  description: string;
  requirements: string;
  /** @nullable */
  benefits?: string | null;
  contactEmail: string;
  status: JobStatus;
  datePosted: string;
  applicationCount?: number;
}

export type JobInputWorkType = typeof JobInputWorkType[keyof typeof JobInputWorkType];


export const JobInputWorkType = {
  Remote: 'Remote',
  Hybrid: 'Hybrid',
  'On-site': 'On-site',
} as const;

export interface JobInput {
  /** @minLength 1 */
  companyName: string;
  /** @nullable */
  companyWebsite?: string | null;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  location: string;
  workType: JobInputWorkType;
  /** @nullable */
  salary?: string | null;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  requirements: string;
  /** @nullable */
  benefits?: string | null;
  contactEmail: string;
  /** @nullable */
  companyContactEmail?: string | null;
}

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  Pending: 'Pending',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;

export type JobUpdateWorkType = typeof JobUpdateWorkType[keyof typeof JobUpdateWorkType];


export const JobUpdateWorkType = {
  Remote: 'Remote',
  Hybrid: 'Hybrid',
  'On-site': 'On-site',
} as const;

export interface JobUpdate {
  status?: JobUpdateStatus;
  title?: string;
  location?: string;
  workType?: JobUpdateWorkType;
  /** @nullable */
  salary?: string | null;
  description?: string;
  requirements?: string;
  /** @nullable */
  benefits?: string | null;
}

export interface Application {
  id: number;
  jobId: number;
  jobTitle: string;
  companyName: string;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  linkedinUrl: string | null;
  /** @nullable */
  cvFileName?: string | null;
  /** @nullable */
  message: string | null;
  dateApplied: string;
}

export interface ApplicationInput {
  jobId: number;
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  cvFileName?: string | null;
  /** @nullable */
  message?: string | null;
}

export interface AdminSummary {
  totalJobs: number;
  pendingJobs: number;
  approvedJobs: number;
  rejectedJobs: number;
  totalApplications: number;
}

export type ListJobsParams = {
search?: string;
location?: string;
workType?: ListJobsWorkType;
status?: ListJobsStatus;
};

export type ListJobsWorkType = typeof ListJobsWorkType[keyof typeof ListJobsWorkType];


export const ListJobsWorkType = {
  Remote: 'Remote',
  Hybrid: 'Hybrid',
  'On-site': 'On-site',
} as const;

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  Pending: 'Pending',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;